rename database_migration directory to database
